requirements.txt
# Core Python Libraries
python-dotenv>=0.19.0

# Blockchain and Web3
web3>=5.30.0
eth-account>=0.5.9

# Exchange APIs
ccxt>=2.0.0

# Async Programming
asyncio>=3.4.3

# Logging and Monitoring
structlog>=21.1.0
prometheus-client>=0.12.0

# HTTP and GraphQL Clients
requests>=2.26.0
gql>=3.4.0

# Data Processing
pandas>=1.3.0
numpy>=1.21.0

# Cryptography and Security
cryptography>=3.4.0

# Error Handling
tenacity>=8.0.0

# Optional: MEV Protection
flashbots>=0.4.0

# Development and Testing
pytest>=6.2.0
pytest-asyncio>=0.15.0
mypy>=0.910
black>=21.5b1
flake8>=3.9.0

# Deployment and Environment
gunicorn>=20.1.0
uvicorn>=0.15.0

# Monitoring and Logging
sentry-sdk>=1.3.0

# Optional: Advanced Trading Features
ta-lib>=0.4.20